# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Odoo Tickets API",
    "description": "\nTicket intake for an Odoo implementation service.\n\n**Authentication.** Every request carries `Authorization: Bearer <token>`.\nThere are two token types:\n\n* `otk_...` \u2014 a long-lived API key, one per client. It belongs in the client's\n  Odoo backend (`ir.config_parameter`, `sudo()`-read only). Never ship it to a\n  browser: anyone with it can file, read, and edit that client's tickets.\n* `ott_...` \u2014 a single-use upload token minted via `POST /api/v1/upload-tokens`.\n  Short-lived, valid only to create one ticket, and pinned to one reporter. This\n  is what you hand a browser so a screenshot can be posted directly.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/whoami": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Whoami",
        "description": "Verify a credential. Use this behind a 'Test connection' button.",
        "operationId": "whoami_api_v1_whoami_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhoAmI"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Credential lacks the required scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "No such resource for this client",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "413": {
            "description": "Request body or attachments too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Field-level validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded; see the Retry-After header",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/upload-tokens": {
      "post": {
        "tags": [
          "tickets"
        ],
        "summary": "Create Upload Token",
        "description": "Mint a single-use token so a browser can upload without the API key.\n\nCall this from the client's Odoo backend, where the API key lives, and hand\nthe returned token to the browser. The reporter identity is baked into the\ntoken, so the browser cannot file a ticket under anyone else's name.",
        "operationId": "create_upload_token_api_v1_upload_tokens_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadTokenOut"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Credential lacks the required scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "No such resource for this client",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "413": {
            "description": "Request body or attachments too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Field-level validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded; see the Retry-After header",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tickets/{ticket_id}": {
      "get": {
        "tags": [
          "tickets"
        ],
        "summary": "Get Ticket",
        "description": "Fetch one ticket by id or by human reference (e.g. `ACME-0042`).\n\nInternal operator notes are never included.",
        "operationId": "get_ticket_api_v1_tickets__ticket_id__get",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketOut"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Credential lacks the required scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "No such resource for this client",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "413": {
            "description": "Request body or attachments too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Field-level validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded; see the Retry-After header",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Seconds to wait"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "tickets"
        ],
        "summary": "Update Ticket",
        "description": "Edit a ticket. Clients may reopen or close, but not mark resolved.",
        "operationId": "update_ticket_api_v1_tickets__ticket_id__patch",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketOut"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Credential lacks the required scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "No such resource for this client",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "413": {
            "description": "Request body or attachments too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Field-level validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded; see the Retry-After header",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Seconds to wait"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorOut": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "detail": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorOut"
      },
...
      "UploadTokenOut": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "max_file_bytes": {
            "type": "integer",
            "title": "Max File Bytes"
          },
          "max_files": {
            "type": "integer",
            "title": "Max Files"
          }
        },
        "type": "object",
        "required": [
          "token",
          "expires_at",
          "max_file_bytes",
          "max_files"
        ],
        "title": "UploadTokenOut"
      },
      "UploadTokenRequest": {
        "properties": {
          "reporter": {
            "$ref": "#/components/schemas/Reporter"
          },
          "ttl_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 3600.0,
                "minimum": 30.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Ttl Seconds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reporter"
        ],
        "title": "UploadTokenRequest"
      },
      "WhoAmI": {
        "properties": {
          "client_id": {
            "type": "string",
            "title": "Client Id"
          },
          "client_slug": {
            "type": "string",
            "title": "Client Slug"
          },
          "client_name": {
            "type": "string",
            "title": "Client Name"
          },
          "api_key_id": {
            "type": "string",
            "title": "Api Key Id"
          },
          "api_key_name": {
            "type": "string",
            "title": "Api Key Name"
          },
          "auth_type": {
            "type": "string",
            "enum": [
              "api_key",
              "upload_token"
            ],
            "title": "Auth Type"
          },
          "server_time": {
            "type": "string",
            "format": "date-time",
            "title": "Server Time"
          }
        },
        "type": "object",
        "required": [
          "client_id",
          "client_slug",
          "client_name",
          "api_key_id",
          "api_key_name",
          "auth_type",
          "server_time"
        ],
        "title": "WhoAmI"
      },
...
    }
  },
...
}